+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+- Introduce primary loader through `#define_primary_loader`.
+- Deprecate `#bulk_load_and_compute` in favor of `#bulk_list_and_compute`.
+
 ## 0.1.1
 
 - Expand docs.